fix featured titles sizing
When the title and/or the subtitle where from the begining larger than
their containers width, its respective col adjustment wouldn't run because
of the width comparison inside the `for` conditional.
The check was moved inside the for declaration to break as necessary
without stoping the column adjusment to run even if the title is wider.